backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.3.1
pydantic==2.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
python-dotenv==1.0.0
aiofiles==23.2.1
pillow==10.0.0
jinja2==3.1.2
fastapi-mail==1.4.1